move h5 methods to preprocess.py and add handling
